use BigInt nanos for storing Durations in Ebean

+++ conf/db/migration/metrics_portal_ddl/V18__report_add_interactive_web_columns.sql
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 
-ALTER TABLE portal.report_sources ADD COLUMN timeout DECIMAL;
+ALTER TABLE portal.report_sources ADD COLUMN timeout BIGINT;
 ALTER TABLE portal.report_sources ADD COLUMN js_run_on_load VARCHAR(8191);
